Featurenico

+++ backend/src/main/java/com/example/metalTest/mantenimientoCorrectivo/controller/request/MantenimientoCorrectivoRequest.java
@@ -3,7 +3,6 @@ package com.example.metalTest.mantenimientoCorrectivo.controller.request;
 import com.example.metalTest.common.validator.ValidEntity;
 import com.example.metalTest.maquina.repository.MaquinaRepository;
 import com.example.metalTest.ordenestrabajo.repository.OrdenesTrabajoRepository;
-import com.example.metalTest.sector.repository.SectorRepository;
 import com.example.metalTest.usuario.repository.UsuarioRepository;
 import lombok.Getter;
 import lombok.Setter;
@@ -30,7 +29,7 @@ public class MantenimientoCorrectivoRequest {
     @NotNull
     @Min(1)
     @Max(2)
-    private int tipofalla;
+    private short tipofalla;
 
     @NotNull
     private int horasProduccionAfectadas;
